requirements.txt
ansible~>1.9.0
ecdsa>=0.13
Jinja2>=2.8
MarkupSafe>=0.23
paramiko>=1.15.2
pycrypto>=2.6.1
PyYAML>=3.11
wheel>=0.24.0
GitPython>=0.3.2.RC1
